# openapi.yaml
openapi: 3.0.3
info:
  title: Área de Membros — Thiago Cantalovo
  version: "1.0.0"
  description: >
    API REST para integrações externas (n8n, Make, Zapier, Flowise, Langflow, checkouts, etc.).
    Documentação em prosa equivalente: API.md. Mantida manualmente — não é gerada a partir do código.
servers:
  - url: https://membros.thiagocantalovo.com
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/webhook/kiwify:
    post:
      tags: [Webhooks de Entrada]
      summary: Notificação de pagamento da Kiwify
      security: []
      parameters:
        - name: signature
          in: query
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: Payload da Kiwify — formato varia por versão, veja API.md
      responses:
        "200":
          description: Sempre 200 — falhas de processamento ficam em webhook_logs, não na resposta HTTP
        "401": { $ref: "#/components/responses/Unauthorized" }
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
...
  responses:
    Unauthorized:
      description: x-api-key ausente ou inválida
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
...